refactor: move serverErrorHandler to apps->www->utils
